+++ packages/firestore/src/core/component_provider.ts
@@ -18,6 +18,10 @@
 import { CredentialsProvider } from '../api/credentials';
 import { User } from '../auth/user';
 import {
+  IndexBackfiller,
+  IndexBackfillerScheduler
+} from '../local/index_backfiller';
+import {
   indexedDbStoragePrefix,
   IndexedDbPersistence
 } from '../local/indexeddb_persistence';
@@ -29,7 +33,7 @@ import {
   MemoryEagerDelegate,
   MemoryPersistence
 } from '../local/memory_persistence';
-import { GarbageCollectionScheduler, Persistence } from '../local/persistence';
+import { Scheduler, Persistence } from '../local/persistence';
 import { QueryEngine } from '../local/query_engine';
 import {
   ClientId,
@@ -87,7 +91,8 @@ export interface OfflineComponentProvider {
   persistence: Persistence;
   sharedClientState: SharedClientState;
   localStore: LocalStore;
-  gcScheduler: GarbageCollectionScheduler | null;
+  gcScheduler: Scheduler | null;
+  indexBackfillerScheduler: Scheduler | null;
   synchronizeTabs: boolean;
 
   initialize(cfg: ComponentConfiguration): Promise<void>;
@@ -105,7 +110,8 @@ export class MemoryOfflineComponentProvider
   persistence!: Persistence;
   sharedClientState!: SharedClientState;
   localStore!: LocalStore;
-  gcScheduler!: GarbageCollectionScheduler | null;
+  gcScheduler!: Scheduler | null;
+  indexBackfillerScheduler!: Scheduler | null;
   synchronizeTabs = false;
 
   serializer!: JsonProtoSerializer;
@@ -115,13 +121,28 @@ export class MemoryOfflineComponentProvider
     this.sharedClientState = this.createSharedClientState(cfg);
     this.persistence = this.createPersistence(cfg);
     await this.persistence.start();
-    this.gcScheduler = this.createGarbageCollectionScheduler(cfg);
     this.localStore = this.createLocalStore(cfg);
+    this.gcScheduler = this.createGarbageCollectionScheduler(
+      cfg,
+      this.localStore
+    );
+    this.indexBackfillerScheduler = this.createIndexBackfillerScheduler(
+      cfg,
+      this.localStore
+    );
   }
 
   createGarbageCollectionScheduler(
-    cfg: ComponentConfiguration
-  ): GarbageCollectionScheduler | null {
+    cfg: ComponentConfiguration,
+    localStore: LocalStore
+  ): Scheduler | null {
+    return null;
+  }
+
+  createIndexBackfillerScheduler(
+    cfg: ComponentConfiguration,
+    localStore: LocalStore
+  ): Scheduler | null {
     return null;
   }
 
@@ -158,7 +179,8 @@ export class IndexedDbOfflineComponentProvider extends MemoryOfflineComponentPro
   persistence!: IndexedDbPersistence;
   sharedClientState!: SharedClientState;
   localStore!: LocalStore;
-  gcScheduler!: GarbageCollectionScheduler | null;
+  gcScheduler!: Scheduler | null;
+  indexBackfillerScheduler!: Scheduler | null;
   synchronizeTabs = false;
 
   constructor(
@@ -184,7 +206,13 @@ export class IndexedDbOfflineComponentProvider extends MemoryOfflineComponentPro
     // set it after localStore / remoteStore are started.
     await this.persistence.setPrimaryStateListener(() => {
       if (this.gcScheduler && !this.gcScheduler.started) {
-        this.gcScheduler.start(this.localStore);
+        this.gcScheduler.start();
+      }
+      if (
+        this.indexBackfillerScheduler &&
+        !this.indexBackfillerScheduler.started
+      ) {
+        this.indexBackfillerScheduler.start();
       }
       return Promise.resolve();
     });
@@ -200,11 +228,20 @@ export class IndexedDbOfflineComponentProvider extends MemoryOfflineComponentPro
   }
 
   createGarbageCollectionScheduler(
-    cfg: ComponentConfiguration
-  ): GarbageCollectionScheduler | null {
+    cfg: ComponentConfiguration,
+    localStore: LocalStore
+  ): Scheduler | null {
     const garbageCollector =
       this.persistence.referenceDelegate.garbageCollector;
-    return new LruScheduler(garbageCollector, cfg.asyncQueue);
+    return new LruScheduler(garbageCollector, cfg.asyncQueue, localStore);
+  }
+
+  createIndexBackfillerScheduler(
+    cfg: ComponentConfiguration,
+    localStore: LocalStore
+  ): Scheduler | null {
+    const indexBackfiller = new IndexBackfiller(localStore, this.persistence);
+    return new IndexBackfillerScheduler(cfg.asyncQueue, indexBackfiller);
   }
 
   createPersistence(cfg: ComponentConfiguration): IndexedDbPersistence {
@@ -283,9 +320,16 @@ export class MultiTabOfflineComponentProvider extends IndexedDbOfflineComponentP
       );
       if (this.gcScheduler) {
         if (isPrimary && !this.gcScheduler.started) {
-          this.gcScheduler.start(this.localStore);
+          this.gcScheduler.start();
         } else if (!isPrimary) {
           this.gcScheduler.stop();
+        }
+      }
+      if (this.indexBackfillerScheduler) {
+        if (isPrimary && !this.indexBackfillerScheduler.started) {
+          this.indexBackfillerScheduler.start();
+        } else if (!isPrimary) {
+          this.indexBackfillerScheduler.stop();
         }
       }
     });